Ticket 48984 - Add lib389 paths module
Bug Description:  lib389 needs a way to consume the paths discovered by
defaults.inf. This allows us to consume that in the new installer, and in
test cases provided by lib389

Fix Description:  Add the paths module that searches wellknown locations for
defaults.inf. If not found, we error. At the moment, this assumes all installs
will follow the defaults.inf, and in the future we will add support for this
to read from dse.ldif if an instance is specified.

https://fedorahosted.org/389/ticket/48984

Author: wibrown

Review by: spichugi, mreynolds (thanks)