Ticket 48991 - Fix lib389 spec for python2 and python3
Bug Description:  Python3 is soon to be the default in fedora. We should be
ready for this, and lay the foundation for rhel adoption of python3 also

Fix Description:  Fix the spec file to create python2-lib389 and python3-lib389
on rhel we only generate the python2 version. Additionally, we gate our new
cli tools to be python3 only in these builds.

https://fedorahosted.org/389/ticket/48991

Author: wibrown

Review by: mreynolds (Thanks!)