Ticket 48983 - Configure and Makefile.in from new default paths work.
Bug Description:   At configure time, this is the only time we know all the
resolved paths of a ds installation. However, external tools such as lib389
need to be able to discover and use these paths.

Fix Description:  This updates defaults.inf with some extra paths that were
found to be needed as part of the lib389 tests.

https://fedorahosted.org/389/ticket/48983

Author: wibrown

Review by: tbordaz, mreynolds (Thanks!)