Ticket 48984 - Add lib389 paths module
Bug Description:  Lib389 needs a way to determine paths for DS when on other
platforms, or prefixed installs. This consumes the new defaults.inf that was
added in #48983

Fix Description:  Use the new ds_paths apis from the instances. This fixes a
large number of tests that would fail on prefixed installs (ie /opt/dirsrv).
This means new tests should be more flexible and able to run on various
platforms. Additionally, we remove the installation prefix option, as it now
conflicts with the new paths module.

https://fedorahosted.org/389/ticket/48984

Author: wibrown

Review by: tbordaz, mreynolds (Thanks!)