Ticket 48982 - Enabling a plugin that has a versioned so causes overflow
Bug Description:  Enabling a plugin that has a versioned.so causes overflow.
This is becuase we assumed that all plugins are "libname.so", and were not
symlinks. So we used a fixed size buffer to realpath.

Fix Description:  Realpath can dynamically allocate the correct size buffer
for the resolved path, so we use that. Additionally, we need to use "free" instead
because realpath uses malloc not slapi_ch_malloc.

https://fedorahosted.org/389/ticket/48982

Author: wibrown

Reviewed by: nhosoi (Thanks so much!)