Ticket 48996 - update DS for ns 0.2.0
Bug Description:  nunc stans changes a number of API's, but most importantly
it removes a number of options.

Fix Description:  Make it possible for DS to use NS again. Remove options that
go away, and add the memalign api which NS now relies on. This removes the
(broken and nonfunctional) enable and disable listener function. It also fixes
the logging of FD exhaustion correctly, so we can see when the server is
overwhelmed. Tested with ldclt.

https://fedorahosted.org/389/ticket/48996

Author: wibrown

Review by: nhosoi, mreynolds (Thanks!)