Ticket 49005 - Update 1.2.11 for lib389 container changes.
Bug Description:  lib389 and DS does not correctly detect or operate in a
container environment yet. We have made the changes in lib389, but this means
we need some extra data in 1.3.5.

Fix Description:  Add the with_systemd flags to the paths defaults.inf so that
lib389 can consume it. Because this is version 1.2.11, this will always be
false.

https://fedorahosted.org/389/ticket/49005

Author: wibrown

Review by: mreynolds (Thanks!)