Ticket 49013 - Correct signal handling with NS in DS
Bug Description:  It was noticed in some tests that we were not correctly
shutting down with nunc stans enabled.

Fix Description:  The issue has a few parts.

First, we aren't setting up the thread pool early enough, so we aren't intercepting
signals early enough. We move the thread pool start to earlier to match existing
nunc-stans-less code.

Second, the signal jobs were not persistent, so they would be consumed after
a single event. If someone hits ctrl-c a bunch this would force kill the server.

Third, we were not handling all the signals we should. We now handle USR1, USR2,
INT, TSTP, TERM and HUP.

Finally, the root cause is an issue in nunc-stans: to resolve this completely
will involve an update to NS 0.2.1.

https://fedorahosted.org/389/ticket/49013

Author: wibrown

Review by: nhosoi, mreynolds (Thanks!)