Ticket 49022 - Lib389, py3 installer cannot create entries in backend
Bug Description:  The issue was that when we created the backend we were not
making the associated mapping tree.

Fix Description:  This fixes addresses a few issues. First, it addes the new
mapping Tree mapped object type, and renames the old one to legacy.

We add the integration between the backend and the mapping tree for creation
and deletion of the mt with the event.

We add a linting event and command to detect invalid backend configurations.

https://fedorahosted.org/389/ticket/49022

Author: wibrown

Review by: spichugi (Thanks!)