Ticket 48707 - Implement draft-wibrown-ldapssotoken-01
Bug Description:  For our new webserver, we need a way to authenticate to the
directory server, while leaving the python webapplication minimal and without
security risk. To achieve this, we will allow DS to implement all security
controls, and the python rest server just passes data back and forth.

Fix Description:  This update allows lib389 to test the ldapssotoken module.
It allows assertion of properties on the rootdse, it adds the plugin handlers
so that we can create and enable the plugin, we add sasl plain support for
testing.

https://tools.ietf.org/html/draft-wibrown-ldapssotoken-01
https://fedorahosted.org/389/ticket/48707

Author: wibrown

Review by: spichugi (Thanks!)