Ticket 49020 - do not treat missing csn as fatal
This patch removes the automatic choice of an alternative csn when
the calculated anchor csn is not found.

In that case it does no longer go to fatal state but will retry later.

It also adds a configuration parameter to thr replication agreement to
allow to pick a "next best" anchorcsn if the original is not found to
keep replicatio going.

Reviewed by: Noriko, William