Ticket 49042 - Increase cache defaults slightly
Bug Description:  We have very small defaults for most sites right now. If we
increase these, we will see a benefit to performance and reduction in cache
eviction. It will help freeipa with it's out of box performance issue before
we move to the full automatic tuning.

Fix Description:  Change the defaults:

    dbcachesize 10MB -> 32MB
    entrycachesize 10MB -> 32MB
    dncachesize 10MB -> 16MB

https://fedorahosted.org/389/ticket/49042

Author: wibrown

Review by: ???