Ticket 48894 - Issues with delete of entrywsi with large entries.
Bug Description:  During delets of entries with a high number of attributes,
a high amount of latency was observed. This was because during the valueset
purge for older csn's the vs->sorted was disposed of. This triggered a rebuild
of the valueset.

Fix Description:  During the delete, don't purge the vs->sorted array. We need
to preserve it if possible which prevents the need to resort the array after
the delete.

https://fedorahosted.org/389/ticket/48894

Author: Mohammad N., Burk, J.

Review by: wibrown, mreynolds (Thanks)