Ticket 49066 - Memory leaks in server - part 2
Bug Description:  This resolves a number of memory leaks and code cleanups in
the server.

Fix Description:  This fixes leaks from server shutdown especially in ldbm
which was not freeing a number of mutexes correctly, and replication which
was not freeing configurations and certain changelog items correctly.

This also resolves a schema issue in plugin configuration creation, and a
segfault in server shutdown when replication has been disabled.

https://fedorahosted.org/389/ticket/49066

Author: wibrown

Review by: mreynolds, tbordaz (Thanks!)