Ticket 49089 - Fix invalid cxxlink statement from hpux
Bug Description:  As a remnant of past build issues, in certain plugins we
defined that we needed a CXXLINK statement. This increased the likelihood of
a broken build in parallel, but does not completely remove it.

Fix Description:  The CXXLINK is not needed now, so we have commented it out
for now, with intent to remove in the future if no issues arise.

https://fedorahosted.org/389/ticket/49089

Author: wibrown

Review by: nhosoi, lslebodn (Thanks!)