Ticket 49089 - List library build deps
Bug Description:  lslebodn notified us of:
https://bugzilla.redhat.com/show_bug.cgi?id=1260190
https://lists.gnu.org/archive/html/automake/2010-03/msg00039.html

This highlighted to me that we were not correctly tracking the dependencies
of our build, where libslapd.la was not a dependency of the make targets
for other plugins. It's better to be explicit!

Fix Description:  Add explicit dependencies lines to our makefile: That way
there can be no mistakes during a compliation.

https://fedorahosted.org/389/ticket/49089

Author: wibrown

Review by: mreynolds (Thanks)