Ticket #49088 - 389-ds-base rpm postinstall script bugs
Description: The uid and gid of the user dirsrv - the owner of ns-slapd are
(389, 389) if the id is available.  If not, the id was increased by one until
unused id was found.  But there is no such requirement.

This patch follows the Fedora Project UID GID soft static allocation strategy.
https://fedoraproject.org/wiki/Packaging:UsersAndGroups#Soft_static_allocation

Reviewed by William <wibrown@redhat.com> and Frank <fhirtz@redhat.com>. (Thanks!!)

(cherry picked from commit c969a8214d591500ba5ec36e54811293b7ebe7d9)