Ticket 47973 - custom schema is registered in small caps after schema reload
Description: The fix for ticket 47973 [1] brought an side effect which
lowers the case of the reloaded new schema.  This patch fixes the issue
by eliminating normalize_oc_nolock().

Note that the normalization is not needed since all the checks are done
by strcasecmp.  There is an exception.  When trying to replace a schema
with the same one having the different cases, exact string compare is
called.

[1] commit e0c78e1c1db351ea5cf835cdb55437b1ea26a3a6

(cherry picked from commit bbd84b451fc015f28658ecbf8c186119cc2cea0d)