Ticket #49104 - dbscan-bin crashing due to a segmentation fault
Description: There was a logic error in format_raw.  When a truncate
option (-t width) is given, the function cut the output and replace
the last 5 bytes with " ...\0".  The position to start the replace
was not correct in some case.

https://fedorahosted.org/389/ticket/49104

Reviewed by wibrown@redhat.com (Thank you, William!!)

(cherry picked from commit efeb2f6e873df32ce7545fa1fb319806d2108fda)