Ticket 49103 - python 2 support for installer
Bug Description:  The python installer tools were originally targeted at python
3 only. Due to the slow adoption of python 3, and it's various issues, we can
not solely rely on it's presence. As a result we need to support python 2
as well.

Fix Description:  Add support for python 2 in the new installer code. This
was mainly related to a number of assumptions present in the init code, and
the use of set manipulations in arrays that is not possible in python2

https://fedorahosted.org/389/ticket/49103

Author: wibrown

Review by: spichugi (Thanks!)