Ticket 49101 - Python 2 generate example entries
Bug Description:  In order to move to the new python installer, we need feature
parity with the existing perl installer. Currently, we are missing one feature,
which is sample entry creation. We need this for our test cases to run.

However, we also don't want to be stuck on the current set of sample entries.
They suck pretty hard, and the acis are just plain terrible. Yet our tests and
customers may depend on them.

Fix Description:  This contains multiple parts to solve the problem.

First, we add a set of identity management objects for groups, ous, users,
services accounts and more. This will greatly simplify our test code, and what
we can create.

Next, we add a set of managers that can apply version configs or sample entries.

We then wire in the capability for instance.create() to use the python installer,
be it via explicit request or environment variable (PYINSTALL) being set.

This way, we can now use the python installer experimentally during tests. A number
of my tests show this already passes.

The great benefit of using python to create the sample entries and configs, is
that they are programatic, versioned, can be unit tested, refactored, they are
portable, no more text file manipulation. We can also use the same code for *upgrade*
as we use for installs, making our tools and proceedures more consistent and
reliable.

A key point of this is that we will be able to install a 1.3.7 server, but with
1.3.6 defaults, or a 1.4.x server with 1.3.7 settings etc. New installs can be
kept in sync with older replicas. Once they are upgraded, we can then do proper
upgrades with the same code that does installs.

This also paves the way for us to add basic user and object management tools for
those unwilling to go the whole way into an IPA install.

https://fedorahosted.org/389/ticket/49101

Author: wibrown

Review by: spichugi (Thanks)