Ticket 49116 - Pblock usage analytics
Bug Description:  In order to refactor and improve the pblock we need to know
what is using it.

Fix Description:  This provides basic usage tracking so that we can account for
access to get and set of the pblock. This allows us to group usage patterns
so that we can refactor.

https://fedorahosted.org/389/ticket/49116

Author: wibrown

Review by: mreynolds (Thanks!)