Ticket 49086 - public api compatability test for SDN changes.
Bug Description:  In order to resolve the SDN premangaling issue, we need to
assert that as we change the pblock and operation code we do not violate our
api behaviours.

Fix Description:  This adds a set of initial tests to assert the behaviours of
the plugin v3 api via the pblock, and partially via the internal operation
api. With this set of tests, changes to the pblock structures and pointers can
be asserted to "behave the same way", which means we will keep the public api
working for existing plugins.

https://fedorahosted.org/389/ticket/49086

Author: wibrown

Review by: mreynolds (Thanks!)