Ticket 47925 - Move add and delete operation aci checks to be before plugins.
Bug Description:  Add, delete and modify had their aci checks in different orders.
This meant that you had different behaviour in all of them.

Fix Description:  Move the add and delete checks to be before calling plugins to
match the modify behaviour. This should allow certain plugins to do entry
transformations that the user is not able to do, allowing proxy or complex
internal operations to be allowed.

https://fedorahosted.org/389/ticket/47925

Author: wibrown

Review by: ???