Ticket 49086 - SDN premangaling broken after SASL change
Bug Description:   SDN premangaling is broken after pw_verify change.
This was because we move the verification code to pw_verify, that
requires a target_address *and* a target_spec to look up what backend
we are about to select.

Fix Description:  After a pre_bind, reset the target spec.

https://pagure.io/389-ds-base/issue/49086

Author: wibrown

Review by: mreynolds (Thanks!)