Ticket 49139 - Update makefile and rpm for import
Fix Description:  As we just imported two libraries this changes our
makefile and configure process. We need to correct configure the system
for this, and add these targets to the makefile.

Additionally, because sds and nunc-stans support doxygen we add
this to the build to not lose functionality.

https://pagure.io/389-ds-base/issue/49139

Author: wibrown

Review by: mreynolds, vashirov (Thanks!)