Issue 49156 - Clean up test suites dir structure and docstrings
Description: We have a lot of plugin directories in dirsrvtests/tests/suites:
- dynamic_plugins
- [PluginName]_plugin - 27 directories
While we are in the process of creating tickets for the porting sprint,
it would be nice to have an organised structure. As discussed on the
Engineering/QE Sync Up meeting we need to move all of them to 'plugins'
dir.

Also, as a part of test suites cleanup, we should organise test plans in
the docstrings and IDs to an each test case (can be generated with
Python 'import uuid; uuid.uuid4()').
create_test.py script should be fixed too and it should check if the ID
is unique throughout the tests.

https://pagure.io/389-ds-base/issue/49156

Reviewed by: wibrown, amsharma (Thanks!)