Ticket 49154 - Nunc Stans stress should assert it has 95% success rate
Bug Description:  We should assert that the nunc-stans stress test
is able to pass 95% of it's connections during an overload scenario.

Fix Description:  Assert we pass 95% of connections. Additionally, we
were not actually running the tests properly, so fix that. Improve
the the work thread function to be slightly faster by better using our
atomic shutdown check.

https://pagure.io/389-ds-base/issue/49154

Author: wibrown

Review by: vashirov (Thanks!)