Ticket 49165 pw_verify did not handle external auth
Bug Description:  During the change to improve sasl and simple bind,
we externalised the backend selection outside of do_bind. In an
auto_bind scenario however, this mean the be was null, causing the
dn to always be invalidated.

Fix Description:  Add a pw_validate_be_dn function, that correctly
checks if we are anonymous, a real be dn, or rootdn. This then allows
the correct authentication of autobinds.

https://pagure.io/389-ds-base/issue/49165

Author: wibrown

Review by: mreynolds (Thanks!)