Ticket 49172 - Fix test schema files
Bug Description:  This issue has two points. First, is that we were not
handling conditions where schema was empty, so we got a none type back.

Second, schema handling moves in 1.3.6, to the join systemschema dir and
schemadir. We should be reflective of this in our test, which requires and
update to the defaults.inf to list the path.

Fix Description:  Update defaults.inf, and the test to correctly
handle none.

https://pagure.io/389-ds-base/issue/49172

Author: wibrown

Review by: mreynolds (thanks!)