Ticket 49174 - nunc-stans can not use negative timeout
Bug Description:  FreeIPA regularly sets up service accounts with
an nsIdleTimeout of -1. As a result of an issue with NS and libevent
this would cause an instant timeout and disconnect of the service
account.

Fix Description:  Correctly check that jobs are registered to NS.
Add validation to NS for negative timeouts. During the job registration,
we force the timeout to be a valid value.

https://pagure.io/389-ds-base/issue/49174

Author: wibrown

Review by: mreynolds(Thanks!!!)

Signed-off-by: Mark Reynolds <mreynolds@redhat.com>