Ticket 49193 - gcc7 warning fixes
Fix Description:  This is the first pass at fixing a number of warnings
that gcc 7 raises. The focus of this patch is mainly:

* Unused parameters
* String format overflows

The benefit to annonating the unused parameters is that when we clean
the codebase for 1.4.x, we can check what is public / private, what's
in cb's etc, and then change the functions as needed. IE we can fix
all the internal functions. Less arguments to a function does improve
perf because of less items on the stack :)

https://pagure.io/389-ds-base/issue/49193

Author: wibrown

Review by: mreynolds (Thanks!)