Ticket 13 - Add init function to create new domain entries
Bug Description:  When we create a new backend, we were missing a
way to populate the entries for usage.

Fix Description:  Add an initialise command to dsidm which is
capable of creating sample entries based on the version supplied
or defaulting to the highest available version.

https://pagure.io/lib389/issue/13

Author: wibrown

Review by: spichugi (Thanks!)