Issue 22 - Specify a basedn parameter for IDM modules
Description: We need a way to create our IDM users and groups in a
custom suffix. Now it is hard coded to 'ou=people' for users,
'ou=groups' for groups, etc.

Also, it is important to do not break current functionality.
CLI IDM depends on it. The parameter should remain optional,
with a default set to 'ou=people' for users, 'ou=groups' for groups, etc.

Also, fix test module names, so they can be successfully
discovered by pytest.

https://pagure.io/lib389/issue/22

Reviewed by: wibrown (Thanks!)