Ticket 49226 - Memory leak in ldap-agent-bin
Bug Description:  We had a leak "somewhere" in the ldap-agent-bin
code. This wasn't able to be easily traced, due to the output of:

Direct leak of 29 byte(s) in 1 object(s) allocated from:
    #0 0x7f7c696e7880 in malloc (/lib64/libasan.so.4+0xde880)
    #1 0x7f7c691afca4 in ber_memalloc_x (/lib64/liblber-2.4.so.2+0x7ca4)
    #2 0x60400000034f  (<unknown module>)

Fix Description:  The issue was that we were not freeing the attr
and val from ldap_parse_line while parsing our entries.

https://pagure.io/389-ds-base/issue/49226

Author: wibrown

Review by: nhosoi (Thanks for the "coverity" check!)