Ticket 49231 - fix sasl mech handling
Bug Description:  In our sasl code we had two issues. One was that
we did not correctly apply the list of sasl allowed mechs to our
rootdse list in ids_sasl_listmech. The second was that on config
reset, we did not correctly set null to the value.

Fix Description:  Fix the handling of the mech lists to allow
reset, and allow the mech list to be updated properly.

https://pagure.io/389-ds-base/issue/49231

Author: wibrown

Review by: mreynolds (Thanks!)

(cherry picked from commit 639535771d13de86446a361a515713ae15f65722)