Ticket 48985 - Add schema for nested groups to work out of box.
Bug Description:  Previously, nestedGroups didn't work on pure
ds because you needed to add inetUser to a group. As well, the
auto objectClass didn't work as it would be NULL by default.

Fix Description:  Add schema and update defaults for memberof
to work out of the box with no alterations. Make it the default
add objectClass, and because of it's simple nature, it can apply
to groups and users.

https://pagure.io/389-ds-base/issue/48985

Author: wibrown

Review by: mreynolds (Thanks!)