Ticket 49053 - Enable flto for DS
Bug Description:  Link Time Optimisation allows GCC to perform
broader and more substantial improvements to our code. This is
described https://gcc.gnu.org/onlinedocs/gccint/LTO-Overview.html

Fix Description:  To enable LTO, there were a number of errors
exposed on libldaputil related to the use of a callback struct.
However, because of the current definition of INTLDAPU, this
meant that the macros and ldap access vector were never accessed.
This cleans up the struct and remove the macros, which resolves
the issue with LTO building.

https://gcc.gnu.org/onlinedocs/gccint/LTO-Overview.html

Author: wibrown

Review by: mreynolds (Thanks!)