Ticket 28 - Change default objectClasses for users and groups
Bug Description:  Previously we had to use inetUser on users
for memberOf, and no option existed for groups.

Fix Description:  With the addition of nsMemberOf we can support users
and groups with this objectClass. As well, it allows us to remove
inetUser because it's attributes are supplied by the other classes
on the object.

https://pagure.io/lib389/issue/28

Author: wibrown

Review by: mreynolds (Thanks!)