From 89e1b5a2f03f031ff2b9423df8d52f1c66de2566 Mon Sep 17 00:00:00 2001 From: William Brown <firstyear@redhat.com> Date: Mon, 22 May 2017 15:44:26 +1000 Subject: [PATCH] Ticket 48864 - remove config.h from spal header.
Bug Description:  In spal header we required config.h, but descentdant
plugins in other projects may not provide it. As a result we should not
rely on it.

Fix Description:  Remove the config check in spal. Additionally, we
need to be stricter on inttypes.h in configure.ac.

https://pagure.io/389-ds-base/issue/48864

Author: wibrown

Review by: viktor & mreynolds