Ticket 49267 - autosize split of 0 results in dbcache of 0
Bug Description:  autosize split of 0 results in a dbcache of 0. This was
due to a missing bounds check on the value for 0. In theory this could
still be problematic if the value was say 1% ... But hopefully we don't
see that :)

Fix Description:  Add the bounds check.

https://pagure.io/389-ds-base/issue/49267

Author: wibrown

Review by: mreynolds (Thanks!)

(cherry picked from commit 22d4865ea20acb6e6c11aed10d09241b09bb711c)