Ticket 49268 - master branch fails on big endian systems
Bug Description:  We had a number of tests that did not function
on a bigendian system, or a system with different pagesizes

Fix Description:  Correct the tests to set the page size to a known
value. Fix the hashtest to use correct endianness for inputs.

https://pagure.io/389-ds-base/issue/49268

Author: wibrown

Review by: mreynolds, lslebodn (Thanks!)