Ticket 62 - dirsrv offline log
Bug Description:  dirsrv log would use the online ldap server to lookup
the access and error path, but this does not work when instance is
offline.

Fix Description:  Convert this to use the ds_paths module that does
correctly account for this. Improve offline state handling so that
we can check this properly.

https://pagure.io/lib389/issue/62

Author: wibrown

Review by: spichugi (Thanks!)