Ticket 49277 - improve out of box system tuning for EL7
Bug Description:  We asked admins to run dsktune and others tools
but with the move to "pet" and automation, we have no guarantee
an admin will tune their system.

Fix Description:  We should provide sane defaults to systemd
and sysctl that enables DS to perf "great" out of the box. This
is a combination of my sysadmin expirence and our performance
tuning guide.

https://pagure.io/389-ds-base/issue/49277

Author: wibrown

Review by: mreynolds (Thanks!)