Ticket 49284 - resolve crash in memberof when deleting attrs
Bug Description:  We were not reseting the entryScope and
excludedScopes to NULL/count = 0 when we deleted the attributes.
This would cause a heap use after free, which mean invalid data
may have been avaliable to the server causing a later segfault.

Fix Description:  When we do not have an entry scope, guarantee
we reset it to NULL/0.

https://pagure.io/389-ds-base/issue/49284

Author: wibrown

Review by: mreynolds (Thanks!)