Ticket 49289 - Improve result handling from connections with NS
Bug Description:  In debug builds with replication it was noticed that
we aren't closing connections at shutdown. This has always been the case
but now with nunc-stans we are seeing it. To make it worse, in a debug
build this would trigger an assertion that caused the server to
abrt().

Fix Description:  Change the result return type from nunc-stans from
simple "true false",to an enum, that allows us to detect more conditions
and make informed decisions as a caller. In the future we need to close
connections at shutdown, but this requires much more work around the
connection process and has been delayed for 1.4.0.

See also: https://pagure.io/389-ds-base/issue/48184

https://pagure.io/389-ds-base/issue/49289

Author: wibrown

Review by: tbordaz (Thanks!)