Ticket 67 - get attr by type
Bug Description:  In python 3 the change to utf8 has caused us some
difficulty. As a result, to help ease this, we need to be able to retrieve
attributes by types in DS

Fix Description:  In mapped object, add the ability to request types
for searches when we perform them.

https://pagure.io/lib389/issue/67

Author: wibrown

Review by: ilias95, spichugi (Thanks!)