Ticket 49310 - remove sds logging in debug builds
Bug Description:  During a load test we noticed that debug builds
would output a large amount of sds logging. Because this uses
stdout, this locks the print buffer causing a serialisation point.

Fix Description:  Macro the SDS logging to -DSDS_DEBUG which is
only required for working on the library.

https://pagure.io/389-ds-base/issue/49310

Author: wibrown

Review by: mreynolds (Thanks!)