Issue 74 - Advice users to set referint-update-delay to 0
Bug Description: Support for asynchronous referential integrity updates
might be dropped in the future and the referint-update-delay attribute
might become deprecated. We need to warn the users.

Fix Description: Display a warning when the healthcheck command is
issued and advise the user to set this attribute to 0.

https://pagure.io/lib389/issue/74

Author: Ilias95

Review by: wibrown (Thanks!)