Issue 78 - Add exists() method to DSLdapObject
Description: Add an exists() method to DSLdapObject which returns
a boolean indicating if an entry exists or not. This can be useful in
cases such as when we want to check whether an entry exists in order
to create it if needed, or when checking for the completion of a task
(no task entry means that the task have been completed).

https://pagure.io/lib389/issue/78

Author: Ilias95

Review by: wibrown (thank you so much!)