Ticket 49275 - shadow warnings for gcc7 - pass 1
Bug Description:  Shadow variables are where we redeclare a var
of the same name in a child scope. In C this leads to a new var
being created, but often represents a fault of the developer
where they may have intended to use the outer scope.

Fix Description:  Rename inner variables, move some from
globals to stack in functions.

https://pagure.io/389-ds-base/issue/49275

Author: wibrown

Review by: mreynolds (Thanks!)