Issue 49329 - Descriptive error msg for USN cleanup task
Bug Description: Providing either a suffix or backend attribute is necessary
in order for a USN tombstone cleanup task entry to be created, otherwise
the server will not accept the operation. However in this case it just
returns a generic error message which is not helpful for adressing the
issue; "Bad parameter to an ldap routine (-9)".

Fix Description: Provide an additional descriptive message in this case.

https://pagure.io/389-ds-base/issue/49329

Author: Ilias95

Review by: wibrown (thanks for this!)